feat: adding function based retries
This allows users to configure retry functionality with logic rather than arbitrary numbers, they can now provide a function that returns the maximum number of retries and a handler.

The handler will be passed the validationError, the request details and the retryCount.

It can use this information to determine if a test should be retried or not and even implement exponential backoff.

To enable this, some other code and documentation has also been adjusted. This should be backwards compatible though.